BMP done

+++ libft/src/string/ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: kaye <kaye@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/04 16:58:04 by kaye              #+#    #+#             */
-/*   Updated: 2021/01/20 18:54:35 by kaye             ###   ########.fr       */
+/*   Updated: 2021/01/20 18:58:11 by kaye             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ char			*ft_itoa(int n)
 	unsigned int	n_tmp;
 
 	if (n < 0)
-		n_tmp = -n;
+		n_tmp = -(unsigned int)n;
 	else
 		n_tmp = n;
 	len = (n < 0) ? ft_intlen(n_tmp) + 1 : ft_intlen(n_tmp);
